feat: add Cart feature and History of Carts; rename History to History of Scans; baseline profile auto-run; unify UiTestTags
- Cart feature:
  - Introduce cart_api / cart_impl modules
  - Add Cart screen, VM, list & bottom bar components
  - Domain: ProductOfCart, ProductInCart, CartTotals, use cases (GetPagedProductsOfCart, CloseCart, DeleteProductFromCart)
  - DB: Cart, CartItem DBOs, DAO, mappers

- History split:
  - Rename 'History' to 'History of Scans' (api/impl)
  - Add 'History of Carts' (api/impl) with list, headers, previews

- UI & Resources:
  - Move common strings to modules/base/resources
  - Add icons (history_of_scans, history_of_carts, cart)
  - New UI kit components (PriceInputField, WeightInputField, ThirdButton, ProductField)
  - ScannedProductUiModel, ProductOfCartUiModel, mappers

- Baseline profile & tags:
  - Auto-run baselineprofile before assemble/bundle Release
  - Replace hardcoded tags with UiTestTags across Scanner & Histories
  - Update GenerateBaselineProfile to target NAV_HISTORY_OF_SCANS and new list tags

- Core API/Impl:
  - Split Products -> ScannedProducts
  - Add CartRepository API + Impl
  - Mappers for Cart/CartTotals/ProductOfCart
  - Use cases: GetActiveCartTotals, DeleteScannedProduct (renamed), etc.

- App wiring:
  - Navigation destinations & routers updated
  - App scaffolding and bottom bar reflect new modules

+++ modules/core/core_api/src/main/kotlin/com/yanchelenko/piggybank/modules/core/core_api/domain/GetPricePerKgUseCase.kt
@@ -1,5 +1,5 @@
 package com.yanchelenko.piggybank.modules.core.core_api.domain
 
 interface GetPricePerKgUseCase {
-    operator fun invoke(weightGrams: Double, price: Double): Double
+    operator fun invoke(weightGrams: Int, price: Double): Double
 }
